backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
flask-sqlalchemy==3.1.1
flask-migrate==4.0.7
flask-jwt-extended==4.6.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
bcrypt==4.1.3
marshmallow==3.21.3
gunicorn==22.0.0